toy_apps/utils/requirements.txt
psycopg2-binary==2.9.1
fastapi==0.68.1
pandas
uvicorn==0.15.0
SQLAlchemy